// frontend/src/types/tourismContent.ts
export interface RecommendedTourismContent {
  contentId: string
  contentTypeId: string | null
  title: string
  address: string | null
  firstImage: string | null
  firstImage2: string | null
  latitude: number | null
  longitude: number | null
  legalDongCode: string | null
  categoryCode: string | null
}